+++ src/main/kotlin/com/freewill/controller/ProfileController.kt
@@ -1,0 +1,42 @@
+package com.freewill.domain.profile.controller
+
+import com.freewill.domain.profile.dto.message.ProfileResponseMessage
+import com.freewill.domain.profile.dto.param.ProfileUpdateParam
+import com.freewill.domain.profile.dto.request.ProfileUpdateRequest
+import com.freewill.domain.profile.dto.response.ProfileDetailResponse
+import com.freewill.domain.profile.service.ProfileService
+import com.freewill.entity.User
+import com.freewill.common.annotation.AuthorizedUser
+import com.freewill.common.response.ApiResponse
+import com.freewill.common.response.ApiResponse.Companion.createSuccessWithData
+import org.springframework.web.bind.annotation.GetMapping
+import org.springframework.web.bind.annotation.PutMapping
+import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.RequestPart
+import org.springframework.web.bind.annotation.RestController
+import org.springframework.web.multipart.MultipartFile
+
+@RestController
+@RequestMapping("/api/v1/profiles")
+class ProfileController(
+    private val profileService: ProfileService
+) {
+    @GetMapping
+    fun searchProfile(@AuthorizedUser user: User): ApiResponse<ProfileDetailResponse> {
+        return createSuccessWithData(
+            ProfileResponseMessage.SUCCESS_SEARCH_PROFILE.msg,
+            profileService.findByUser(user).toProfileDetailResponse()
+        )
+    }
+
+    @PutMapping
+    fun updateProfile(
+        @AuthorizedUser user: User,
+        @RequestPart(value = "image", required = false) image: MultipartFile?,
+        @RequestPart(value = "profileUpdateRequest", required = false) request: ProfileUpdateRequest?
+    ): ApiResponse<Void> {
+        profileService.update(user, ProfileUpdateParam(image, request?.nickname, request?.email))
+
+        return createSuccessWithData(ProfileResponseMessage.SUCCESS_UPDATE_PROFILE.msg)
+    }
+}
