--- src/main/kotlin/com/freewill/domain/profile/dto/message/ProfileResponseMessage.kt
@@ -1,6 +1,0 @@
-package com.freewill.domain.profile.dto.message
-
-enum class ProfileResponseMessage(val msg: String) {
-    SUCCESS_SEARCH_PROFILE("프로필 조회 성공"),
-    SUCCESS_UPDATE_PROFILE("프로필 수정 성공")
-}
