--- src/main/kotlin/com/freewill/domain/user/dto/message/AuthResponseMessage.kt
@@ -1,6 +1,0 @@
-package com.freewill.domain.user.dto.message
-
-enum class AuthResponseMessage(val msg: String) {
-    SUCCESS_SIGN_UP("회원 가입 성공"),
-    SUCCESS_SIGN_IN("로그인 성공")
-}
