--- src/main/kotlin/com/freewill/domain/user/dto/request/SignInRequest.kt
@@ -1,8 +1,0 @@
-package com.freewill.domain.user.dto.request
-
-import com.freewill.domain.user.entity.enums.Provider
-
-data class SignInRequest(
-    val provider: Provider,
-    val idToken: String
-)
