--- src/main/kotlin/com/freewill/domain/user/repository/UserRepository.kt
@@ -1,10 +1,0 @@
-package com.freewill.domain.user.repository
-
-import com.freewill.domain.user.entity.User
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.stereotype.Repository
-
-@Repository
-interface UserRepository : JpaRepository<User, Long> {
-    fun findByProviderId(providerId: String): User?
-}
