+++ src/main/kotlin/com/freewill/dto/param/OAuth2Param.kt
@@ -1,0 +1,10 @@
+package com.freewill.domain.user.dto.param
+
+import com.freewill.enums.Provider
+
+data class OAuth2Param(
+    val provider: Provider,
+    val providerId: String,
+    val providerNickname: String,
+    val providerEmail: String?
+)
