+++ src/main/kotlin/com/freewill/dto/request/SignInRequest.kt
@@ -1,0 +1,8 @@
+package com.freewill.domain.user.dto.request
+
+import com.freewill.enums.Provider
+
+data class SignInRequest(
+    val provider: Provider,
+    val idToken: String
+)
