+++ src/main/kotlin/com/freewill/s3/S3Uploader.kt
